Removed redundant Stack.Peek call
We can do the `Pop` and test in the same operation. We might find the right value on the stack (the opening paren) in which case we _do_ want to have popped it. Alternatively we'll find some unexpected value in which case we need to return false and it doesn't matter that we've popped something we might not h\ve wanted to!