Pushed list of delimiters to use onto IsBalancedTestData
This removes duplication from the two test fixtures. It also means we're not splitting the definition of the test case across multiple types i.e. having the input string in one type, but the delimiters we should use on another. How can `IsBalancedTestData` possibly say whether the input should be considered balanced if it doesn't know what delimiters are in play?!