Minor refactoring of IsBalanced
The introduction of an extension method makes the test we're doing a little clearer.

We could drop one of the conditionals. If the `value` string is empty it cannot contain any of the delimiter characters so is reasonably handled by the second conditional.

By introducing a variable for the test condition we can better document the semantics of the check.