Used more appropriate loop in implementation of IsBalanced
By virtue of letting the IDE replace tail recursion with a loop we had deliberately set up an infinite loop, but exited it using a return statement. This has now been replaced by a loop with an appropriate exit condition. It also removes the naughtiness of using one of the method parameters as a variable. This had also been introduced when removing the tail recursion. Introducing a variable to use means we can give it a better name, and further help document the method.