Fixed failing test(s) for StringReplacementBalancedCalculator
Using a greedy quantifier (`*`) meant with an unbalanced string such as "\`A\`, \`B\`, \`C" we would match "A \`, \`B\`, " as the delimited text and consider this a match. Changing to a lazy quantifier fixed this scenario as we would then get two matches (the quoted A and B) and leave the unbalanced delimiter behind as expected.

This introduced a different issue if we had nested but balanced delimiters e.g. "[[]]". Because we match left to right this would match "[[]" and cause us to incorrectly think there was an unbalanced closing delimiter. The fix was to also include the "current" delimiter characters in the negated character group!

This means was can actually simplify the code somewhat. We don't have to extract one delimiter from the list when calling `GetPattern` each time around the loop. We can also simplify GetPattern. We can now assume there will _always_ be at least one delimiter (the current one) when building the character group.

Includes moving the quantifier into the string concatenation. I think that makes it clearer what the overall pattern is.