Removed unnecessary building of string of delimiters
We _had_ to do this _inside_ the loop previously because the list of delimiters to include in the character group would always change i.e. _not_ include the current delimiter.

Recent fixes for failing tests mean the string of escaped delimiters will now _always_ be the same; it is for all `Delimiter`s passed to `IsBalanced`. As such we can build the string once then supply that to `GetPattern` rather than give `GetPattern` the `IEnumerable<Delimiter>` and have that do the string building.