Fixed failing tests for StackBasedBalancedCalculator for when the start and end delimiter are the same character.
Because the delimiter characters were the same we would always interpret the character in the input string as the opening delimiter and thus push it onto the stack. That meant when the delimiters were properly paired we would end up with both pushed onto the stack rather than one pushed, and then popped. We'd end up with two items in the stack, rather than none, and assume that meant we had unbalanced delimiters!

The algorithm is now a bit clunky with nested conditionals inside a loop 🙁