Removed nested conditional from StackBasedBalancedCalculator
We always need to do the same thing for a starting character regardless of whether the delimiter represents a matching pair (e.g. double quotes) or an opposing pair (e.g. angle brackets). The same holds true for the ending character i.e. the logic is the same regardless of whether the delimiter is a matching or opposing pair.

The "logic" actually revolves around telling whether we're looking at the start or end character. This conditional has been made more obvious by breaking up the different parts of the expression, and giving them named values.