Added simple test cases for delimiters with length greater than 1.
For now we're just looking at single delimiters and without nesting or repetition.

We've only applied these cases to `StringReplacementBalancedCalculator` which has handled five of the seven cases without modification. That leaves us with a couple of failing tests where the generated regex pattern is not valid. This is down to putting hyphens into a character group which are then being interpreted as a character range rather than text to (not) match.

For example "<%--[^$<%----%>]*?--%>" cannot be parsed as a regex pattern.