Enhanced StackBasedBalancedCalculator to handle multi-character delimiters
There are a couple of funky bits of code in here, but it at least passes all test cases!

We can no longer iterate the input string as a `Char[]` because we (possibly) need to be looking at several characters at a time. That means we have to use an indexer into the string. However, to account for (possibly) needing to skip some characters in the input string we're manipulating the loop variable. That is generally considered bad practice.

The single-line expression to work out which `Delimiter` (if any) we're currently looking at is hard to understand, as well as having a repeated expression in the lambda statement.