Made finding current delimiter easier to read, and removed duplication
We can turn the list of `Delimiter`s into what is essentially a multi-dimensional array of strings. We can then check whether any element in the multi-dimensional array exists at the specified index in the input array. This test was implemented as an extension method of `string` to aid legibility. If we do find a match on any of those strings we know it will be at the same index as the corresponding `Delimiter`. This has also given us the option to test against a null for "not found" rather than the previous (slightly opaque) technique of testing against the default of the data type!