Stopped compiling the regular expression
We only anticipate each `Regex` instance being used a handful of times. Because we don't know the delimiters ahead of time we have to create an instance on each call to `IsBalanced`. We will then only call it a handful of times which is dependent on the length of input string and how many occurrences of delimiters there are. We certainly don't expect to call it often enough to justify the startup cost of compiling it!

If the `RegexBasedBalanceCalculator` was constructed with a set of `Delimiter`s and the same set used on _every_ call to `IsBalanced` it _might_ make sense to compile the `Regex`.

Removing this flag resulted in the unit tests for that type running about 10x faster! 😲 I was previously seeing them take ≈ 200ms. I'm now seeing them take ≈ 20ms.