Added method to find _all_ indices of a given substring in a string
These tests are currently in `PairedIndicesBalancedCalculatorTests` even though they are not testing `PairedIndicesBalancedCalculator`. They are, however, testing parts of an algorithm we intend to "lift and shift" into it.

This is intended to be used as part of a new implementation of `IBalancedCalculator`. The idea is we will find all indices of the delimiters and then pair  up the starts and ends. We will then look for any pairs that don't overlap with any other pairs and remove them. We will repeat this until no more pairs can be removed. If there are no pairs left the input is balanced. If there are any pairs left the input is not balanced.