Added method to find indices of Start and End strings of an instance of Delimiter
These tests are currently in `PairedIndicesBalancedCalculatorTests` even though they are not testing `PairedIndicesBalancedCalculator`. They are, however, testing parts of an algorithm we intend to "lift and shift" into it.

We're returning tuples which I ordinarily would not do. However, this method will end up being private implementation details, rather than on a public interface, so using tuples might be reasonable.