Implemented functionality to pair up indices of start and ending delimiters in an input string
This makes the assumption the two input arrays will only be referring to a single Delimiter instance i.e. they only have the locations of opening and closing parens, but not the locations of opening and closing braces etc.

These tests are currently in `PairedIndicesBalancedCalculatorTests` even though they are not testing `PairedIndicesBalancedCalculator`. They are, however, testing parts of an algorithm we intend to "lift and shift" into it.

We're returning tuples which I ordinarily would not do. However, this method will end up being private implementation details, rather than on a public interface, so using tuples might be reasonable.