Implemented method to any identify the elements of an enumerable of start and end indices that _don't_ overlap with any other elements
These tests are currently in `PairedIndicesBalancedCalculatorTests` even though they are not testing `PairedIndicesBalancedCalculator`. They are, however, testing parts of an algorithm we intend to "lift and shift" into it.

We're returning tuples which I ordinarily would not do. However, this method will end up being private implementation details, rather than on a public interface, so using tuples might be reasonable.

The expression is currently _awful_ and not at all clear what we're trying to do (unless you look at the tests). It does, however, work and we can tidy it up 🙂