Added initial implementation of PairedIndicesBalancedCalculator
So far we've only added in one of the `TheoryData` instances, and we've _mostly_ satisfied that. We currently have three failing tests. All three are related to not correctly identifying delimiters are paired when the start and end characters are the same.

Most of this was implemented by shifting code from the test fixture into the private implementation of the type, with a few minor changes.

Some of the code is a bit smelly, but we'll sort that out when we've got all tests passing.

The most significant of was the introduction of a record to hold a pair of start and end indices. This means we're passing fewer tuples around, so method signatures are clearer.

A second significant change was the introduction of a test when pairing indexes to see if we actually found the corresponding end index for a given start index.