Fixed bug that would cause PairedIndicesBalancedCalculator to think balanced delimiters were unbalanced when the start and end character were the same.
Because the start and end delimiting character the list of indices for the start character is the same as the list of indices for the end character. In other words we were identify characters we _should_ have been treating as end delimiters as starting ones, and vice versa.

We should have been treating the first character and _every other one_ as start characters. Similarly, we should have been treating the second character and every other one as an end character!

The obvious way to implement this was looking at the index of indices we had picked out. If the index was an odd number we should treat it as a start character. If it was even was should treat it as an end character.