Added remaining test cases for PairedIndicesBalancedCalculator
Boom! 💣 Because we were pretty happy with the implementations of the different steps of the algorithm from the exploratory tests we added to the test fixture this handled the more complex scenarios without us having to make further changes 😀

I was _expecting_ this implementation to perform _terribly_. It is actually the fastest of the three. It is a little faster than the stack-based solution, but massively faster than the regex-based solution. Why is it faster than the stack-based solution? I have absolutely no idea!