Minor refactoring... Changed local methods into public methods on nested type
Local methods always feel a bit grotty / clumsy to me. Given the tests in question were only ever about comparing one `IndexPair` to another it seemed that type was the appropriate location for the query method, in exactly the same way that is where you would expect to find e.g. operators such as ==, + etc