Performance optimisation for PairedIndicesBalancedCalculator
By switching to `foreach` loops rather than `Select` we have the option to short-circuit some evaluation.

If we know there is an imbalance on the number of start and end indices on the first delimiter there is absolutely no point in enumerating the input string repeatedly to get the indices for the second, third, fourth etc delimiters.

If we know we didn't pair up the indices for the first delimiter there is no value in pairing the indices for the second, third, fourth etc delimters.