Performance optimisation for and simplification of PairedIndicesBalancedCalculator
Originally we were using a `Select` operation and as a consequence end up holding a list of tuples, and one field of that tuple was the `List<IndexPair>`. We needed the entire tuple so we could check the `Success` field.

When we changed to a `foreach` loop we could check `Success` within the loop and thus not need to store it; we only needed the `List<IndexPair>` from the tuple.

Rather than storing `List<List<IndexPair>>` and then having to collapse it into a single list by the slightly-odd looking `.SelectMany(x => x)` we might as well just store it as a flat list in the first place!

This was done to simplify the code. It was a nice bonus that we get a performance improvement from it as well 😊