Minor simplification of PairedIndicesBalancedCalculator
We only need the delimiter to when we are starting to search for indices. Once we have them we have the link between the list of start indices, and the list of end indices; they are part of the same tuple. As such, we don't need the delimiter any more and can discard it. We could see this was the case because we were using the discard variable!