Added implementation that takes advantage of the fact the pattern of answers repeats every 15 items
As the pattern repeats we can get away with storing a small lookup of how to get the answer for values in the range 0..14, and then figure out where in the repeating pattern the value lies. Armed with that information we can then get a function to calculate the answer for us.  We can't store the _actual_ answer because we'd end up returning e.g. 1 for an input of 16!