Added FizzBuzz implementation that stores values that should Fizz or Buzz in an array
This works on the same principle as the previous implementation ie that the answers for a pattern that repeats every 15 elements. However, rather than storing a `Func<int, string>` that can give us the answer and then use the value to index into it we flip this around; the _value_ is the element in the array, and we check for its existence.