Added optimised implementations for parsing an int from a string
The basic idea of all three implementations it to get the `Char` at a specific index in the input string, convert it into its `ushort` representation of the same number, work out what power of 10 to multiply that by, and sum all of those values.

The first two version both use our optimised version of Math.Pow to get the power of 10 value, with one using two loop variables running in opposite directions, and one calculating it based on the current index and the length of the input string. Both of these are a little faster than Int.Parse

The third was a realisation we don't need to use that operation if we enumerate the string backwards instead of forwards. If we do that we can use simple multiplication to track the power of 10. This version proved to be _much_ faster, and is probably the one to now focus on seeing if we can make even faster!