Further optimisation of code to parse an int from a string
We'd found a fast way of looping through the string. Looking at how we then accessed the data seemed reasonable. Treating the string as a `char[]` rather than turning it into a `ReadOnlySpan<char>` proved quite a bit faster. Having the span passed into the method was (somewhat surprisingly to me) quite a bit slower.