Further optimisation of code to parse an int from a string
This time trying a different expression within the loop to see if we can get that faster. Rather than manually tracking a multiplier for each character in a separate variable we're actually tracking it with the same variable we use to store the return value. This is achieved by multiplying the return value by 10 on each iteration! The result is we cut the number of operations on each iteration from six (addition, subtraction, 2 multiplication, and 2 assignments) to four (addition, subtraction, multiplication, and assignment. This is _much_ faster 😃

Also trying the same expression with a foreach loop rather than a for loop. This was a big pile of fail!