Tried implementation of IndicesOf using ReadOnlySpan instead of strings
This applies to both the incoming parameters, _and_ how we're holding temporary data during the enumeration.

With `IndexOf` not giving us an overload that accepts a starting location (in the way `string.IndexOf` does) we have to keep slicing the span in order to find subsequent occurrences of `substring`. In turn that means we have more work to do to track what the actual index is. The nature of `ReadOnlySpan` means we can't use `yield return` so have to go back to maintaining a list of the indices we have found. These things in combination mean this implementation is only marginally faster than the baseline, and slower than the version using `yield return`. Unsurprisingly we're back to consuming the same memory as the baseline.

Another attempt using `Stack<int>` in the hope the data structure might be more optimised for our use was unsuccessful. That actually ends up slower than the baseline whilst still consuming the same amount of memory.