Attempted to optimise IndicesOf by manually comparing the substring with an index position in the value
We can't use the equality operator here as that only returns true if two spans point to the same memory, _not_ whether different areas in memory contain the same  data! That means we have to write our own (local) method to compare the contents of the two spans. The hope was by needing to track less variables etc we'd end up being faster. We're actually _much_ slower than the baseline 😢

Includes adding a specific test for the same value and substring we use in the benchmarks. The initial version of this implementation was throwing `IndexOutOfRangeException`s when running the benchmarks!