Attempted to improve performance of IndicesOf when comparing two spans
Rather than effectively creating a new span of appropriate length from an index into the original then comparing that to the substring attempted to compare an offset position in the original. The hope was we'd end up with less allocations, checks etc. This _is_ quite a bit faster than the first implementation comparing spans, but still significantly slower than our baseline.

So far a typical approach to this problem, but using `yield return` has been our best option.