Attempted to improve performance of IndicesOf when comparing two spans
The initial idea was to remove the parameters from the nested function, and consume the `value` and `substring` parameters directly. However, the language does not allow us to do that:

> [CS1628] Cannot use ref, out, or in parameter 'value' inside an anonymous method, lambda expression, query expression, or local function

As such we replaced the local function with a nested loop. This _does_ run faster than the version with the local function, but is still substantially slower than the baseline. We also consume more memory and get more garbage collection which I was not expecting!