Rearranged my implementations of FizzBuzz
Pushed each implementation into its own type, and with an instance method that implements an interface. This is far more in keeping with how we'd expect to consume a type in a real application. It also makes it much more obvious which (if any) data is associated with each implementation.

Rearranging this code hasn't shown any significant change in the benchmarks. The implementations that were fast before still are, and the ones that were slow before still are.

All tests were placed in the same file as they're one line and the test(s) for every implementation are the same.