Added implementation of FizzBuzz that works by indexing into a pre-calculated string
I'm surprised this a) runs _significantly faster than what is the same algorithm but using a char array and b) it consumes less memory with less garbage collection. Maybe the char array version is doing some additional allocations when we slice the char array to the passing into the string ctor, but that is guesswork at this stage!