Added implementations of FizzBuzz that index into a dictionary of strings
The two implementations both work based on knowing there are only really 15 answers we need to know, and it is reasonable to keep those in a dictionary. One implementation checks whether a key exists, and it _not_ existing means we need to return the value. The second implementation solves the same problem by storing the nulls and then coalescing.

Both have similar performance and memory footprint. Both are slower than the baseline, and thus slower than our solutions that use pattern matching.