Added faster AddDigits by not using Linq
Took my fastest implementation and removed the use of maintaining a `List<uint>` by maintaining a running total. This meant I did not then have to use a LINQ `Sum` operation. This is _much_ fast, typically only need between 2 and 4% of the execution time of the benchmark! 🤯 Because everything is being done with primitives there is no heap allocation either.