Added AddDigits that doesn't use tail recursion
As suspected, in spite of Rider grumbling this didn't make much difference. Presumably the compiler is optimising the recursion away although we're not looking at maximum stack depth so we don't know.

Includes changing descriptions of the benchmarks so they render in the order of the different implementations I had tried.