Instructed Rider to ignore tail recursion in DigitsAdder
We _know_ about the recursion, and don't much care. We're not seeing better performance by replacing it with a loop. We also know we're going to get at most three frames on the stack. We're also trying different versions of the implementation, some of which _require_ us to use the recursion.