Separated implementations of AddDigits into multiple types
By breaking apart the large class into multiple implementations of the same interface we:

* Reduce boilerplate in the tests
* Reduce boilerplate in the benchmarks
* Can make the tests self-maintaining
* Can make the benchmarks self-maintaining

We arguably lose a little here; it is harder to control which tests amd benchmarks get executed. It is also a little harder to see what the test or benchmark was using. This can be alleviated by overriding ToString on the implementations if we care that much. You could make a case that it is a little harder to see what the `TheoryData` is, but it is clear enough to me.