Added AddDigits that uses regex to get digits
As expected this performed abysmally compared to the other implementations that I had tried. This was added purely to demonstrate another technique.

Although the regex was not compiled considering it was the only one in use it should have been cached after compilation and, as such, pre-compilation _shouldn't_ have made any significant difference.