Added a proper JS implementation of add digits
This replaces hacking around with a file executed via Quokka with a set of Jest tests to prove the implementation works. Some extra behaviour - being able to parse strings - was added as an excuse to add a couple more tests. Some error handling was included also as an excuse to add more tests.

The JavaScript was moved outside of the C# source code. The directory names are a little off because I never anticipated this having "not C#" code in it when I set this repo up. Because we now have one of the two implementations in a more sensible location it was removed from the Quokka file.