Parameterised Jest tests for addDigits when using string
Having previously tried the array API for parameterising the tests this was just a chance to try the table API. It does show us being able to correctly type the input as string. It also shows us being able to use named parameters, rather than positional parameters in the test description.